backend/summarization_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic[email]==2.7.1
requests==2.31.0

# Mongo
pymongo==4.6.0

# Celery + RabbitMQ
# celery==5.2.7

# Transformers Summarization
# transformers==4.41.2
# torch==2.3.0

# Dotenv
python-dotenv==1.0.1

# Testing
pytest==7.4.2
httpx==0.24.1
pytest-cov==4.1.0
